merges two related launch tests
